+++ bitbybit-tests/src/bitfield_tests.rs
@@ -1678,3 +1678,42 @@ fn test_debug_impl() {
     let display_str = format!("{:?}", test);
     assert_eq!(display_str, "Test { upper: 31, lower: 47 }");
 }
+
+#[test]
+fn introspection() {
+    #[bitfield(u32, introspect)]
+    struct Bitfield {
+        #[bits(0..=7)]
+        a: u8,
+        #[bits(8..=15)]
+        b: [u8; 2],
+        #[bits([24..=25, 30..=31])]
+        c: u4,
+        #[bit(29)]
+        d: bool,
+        #[bit(30)]
+        r#ref: [bool; 2],
+    }
+
+    // <NAME>_BITS exposes the "bits" value(s):
+    assert_eq!(Bitfield::A_BITS, 0..=7);
+    assert_eq!(Bitfield::B_BITS, 8..=15);
+    assert_eq!(Bitfield::C_BITS, [24..=25, 30..=31]);
+    assert_eq!(Bitfield::D_BITS, 29..=29);
+    assert_eq!(Bitfield::REF_BITS, 30..=30);
+
+    // Arrays also have <NAME>_COUNT and <NAME>_STRIDE
+    assert_eq!(Bitfield::B_COUNT, 2);
+    assert_eq!(Bitfield::B_STRIDE, 8);
+    assert_eq!(Bitfield::REF_COUNT, 2);
+    assert_eq!(Bitfield::REF_STRIDE, 1);
+
+    // <name>_mask() returns a mask for the field
+    assert_eq!(Bitfield::a_mask(), 0x000000FF);
+    assert_eq!(Bitfield::b_mask(0), 0x0000FF00);
+    assert_eq!(Bitfield::b_mask(1), 0x00FF0000);
+    assert_eq!(Bitfield::c_mask(), 0xC3000000);
+    assert_eq!(Bitfield::d_mask(), 0x20000000);
+    assert_eq!(Bitfield::ref_mask(0), 0x40000000);
+    assert_eq!(Bitfield::ref_mask(1), 0x80000000);
+}